KeyCollector Refactor
Refactored the key collector and the resource manager. Also split the
resource manager so that a more lean version can be used for network
applications. Creatable IDs are now used rather than ushorts.

+++ PlasmaCore/KeyedObject/KeySeed.cs
@@ -6,9 +6,9 @@ using System.Text;
 namespace Plasma {
     public static class plKeySeed {
 
-        static plKey fAvatarMgr = new plKey((ushort)plCreatableID.plAvatarMgr, "kAvatarMgr_KEY");
-        static plKey fClient    = new plKey((ushort)plCreatableID.plClient, "kClient_KEY");
-        static plKey fNetCliMgr = new plKey((ushort)plCreatableID.plNetClientMgr, "kNetClientMgr_KEY");
+        static plKey fAvatarMgr = new plKey(plCreatableID.plAvatarMgr, "kAvatarMgr_KEY");
+        static plKey fClient    = new plKey(plCreatableID.plClient, "kClient_KEY");
+        static plKey fNetCliMgr = new plKey(plCreatableID.plNetClientMgr, "kNetClientMgr_KEY");
 
         public static plKey AvatarMgr {
             get { return fAvatarMgr; }
